20220527spring

+++ src/main/java/com/choong/spr/controller/BoardController.java
@@ -1,8 +1,11 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -39,7 +42,12 @@ public class BoardController {
 	}
 	
 	@PostMapping("insert")
-	public String insert(BoardDto board, RedirectAttributes rttr) {
+	public String insert(BoardDto board,Principal principal, RedirectAttributes rttr) {
+		
+		System.out.println(principal);
+		System.out.println(principal.getName()); // user name
+		
+		board.setMemberId(principal.getName());
 		boolean success = service.insertBoard(board);
 		
 		if (success) {
@@ -56,35 +64,57 @@ public class BoardController {
 		BoardDto dto = service.getBoardById(id);
 		List<ReplyDto> replyList = replyService.getReplyByBoardId(id);
 		model.addAttribute("board", dto);
-		model.addAttribute("replyList", replyList);
+		// ajax로 처리 
+		// model.addAttribute("replyList", replyList);
 		
 	}
 	
 	@PostMapping("modify")
-	public String modify(BoardDto dto, RedirectAttributes rttr) {
-		boolean success = service.updateBoard(dto);
+	public String modify(BoardDto dto, RedirectAttributes rttr, Principal principal) {
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+		// 게시물 정보 얻고 
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		// 게시물 작성자(memberId)와 principal의 name과 비교해서 같을 때만 진행.
+		if(oldBoard.getMemberId().equals(principal.getName())) {
+			
+			boolean success = service.updateBoard(dto);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			}
+
 		} else {
-			rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
 		}
 		
 		rttr.addAttribute("id", dto.getId());
-		
 		return "redirect:/board/get";
 	}
 	
 	@PostMapping("remove")
-	public String remove(BoardDto dto, RedirectAttributes rttr) {
+	public String remove(BoardDto dto, Principal principal, RedirectAttributes rttr) {
 		
-		boolean success = service.deleteBoard(dto.getId());
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+		// 게시물 정보 얻고 
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		// 게시물 작성자(memberId)와 principal의 name과 비교해서 같을 때만 진행.
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.deleteBoard(dto.getId());
 			
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+				
+			} else {
+				rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			
+		}
+		// 아니면 리턴
+		
 		} else {
-			rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
+			rttr.addAttribute("id", dto.getId());
+			return "redirect:/board/get";
 		}
 		
 		return "redirect:/board/list";
